requirements.txt
discord.py==2.5.2
python-dotenv==1.0.0
motor==3.3.1
dnspython==2.4.2
asyncio==3.4.3